properly set gid
three things were wrong:
1.  failing to check $! after system command (so failure was ignored)
2.  must set gid BEFORE setting id (otherwise you get operation not permitted)
3.  need to set effective id to a list of all groups user is a member of

number 3 was the biggest pain to fix.  without it, however, the user wouldn't
be added to supplimental groups when they were inside the chroot and they would
incorrectly be a member of the root group.

It's a huge PITA to call getgroups( $user ), (getting the groups for the
current user is easy, but an arbitrary user is hard).  So I just
iterated over all the groups to determine which groups user is a member
of.